csc-ml/requirements.txt
Flask==1.1.4
Werkzeug==1.0.1
markupsafe==2.0.1
openai
chromadb
python-dotenv
langchain
bs4
tiktoken
flask-cors